+++ runtime/runtime.cmake
@@ -1,6 +1,7 @@
 prepend(KPHP_RUNTIME_DATETIME_SOURCES datetime/
         datetime.cpp
         datetime_functions.cpp
+        datetime_zone.cpp
         timelib_wrapper.cpp)
 
 prepend(KPHP_RUNTIME_MEMORY_RESOURCE_SOURCES memory_resource/
